// gdr-web-version-mdp/gdr-web-version-mdp/src/app/@data/model/maestra-parametro.ts
export class MaestraParametro {
  parametroId: number;
  tipoParametro: string;
  codigoTexto?: string;
  codigoNumero?: number;
  valorTexto?: string;
  valorNumero?: number;
  valorFecha?: any;
  descripcion?: string;
  estadoRegistro?: string;
  usuarioCreacion?: string;
  fechaCreacion?: string;
  usuarioModificacion?: string;
  fechaModificacion?: string;

  responsableId: number;
  descripcionResponsable?: string;
}